Split pdbrc module
