+++ ansible_collections/scruffaluff/bootware/roles/python/files/pdbrc
@@ -14,6 +14,9 @@ import pdbrc
 pdbrc.setup(pdb.Pdb)
 sys.excepthook = pdbrc.break_exception(pdb)
 
+# Load debug helper modules.
+import plotrc
+
 # Remove setup import modules.
 _ = sys.path.pop()
 if "os" not in _namespace: del os
