+++ ansible_collections/scruffaluff/bootware/roles/python/tasks/config.yaml
@@ -187,10 +187,12 @@
   loop:
     - dest: .pdbrc
       src: pdbrc
+    - dest: .config/pyrc/dbgrc.py
+      src: dbgrc.py
     - dest: .config/pyrc/pdbrc.py
       src: pdbrc.py
-    - dest: .config/pyrc/sitecustomize.py
-      src: sitecustomize.py
+    - dest: .config/pyrc/plotrc.py
+      src: plotrc.py
   tags:
     - config
   when: ansible_facts.system != "Win32NT"
@@ -203,10 +205,12 @@
   loop:
     - dest: .pdbrc
       src: pdbrc
+    - dest: .config\pyrc\dbgrc.py
+      src: dbgrc.py
     - dest: .config\pyrc\pdbrc.py
       src: pdbrc.py
-    - dest: .config\pyrc\sitecustomize.py
-      src: sitecustomize.py
+    - dest: .config\pyrc\plotrc.py
+      src: plotrc.py
   tags:
     - config
   when: ansible_facts.system == "Win32NT"
